tests/wpt/web-platform-tests/web-nfc/NDEFReader-document-hidden-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta name="timeout" content="long">
<title>NDEFReader.onreading respect page visibility changes</title>
<link rel="help" href="https://w3c.github.io/web-nfc/#visible-document">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

promise_test(async t => {
  const reader = new NDEFReader();
  reader.scan();
  const readerWatcher = new EventWatcher(t, reader, ["reading", "error"]);

  return await new Promise((resolve, reject) => {
    readerWatcher.wait_for("reading").then(event => {
      if (document.hidden) reject();
      resolve();
    });
  });
}, "Test NDEFReader.onreading is not fired when document is hidden");

</script>

<p>Step1: switch the page to the background, then tap a formatted NFC tag.</p>
<p>Step2: switch back to the page, then tap the tag again.</p>
